requirements.txt
numpy==1.25
pandas==2.0
pyahocorasick==2.0
